backend/requirements.txt
alembic==1.12.1
click==8.1.7
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-session==0.6.0
greenlet==3.0.3
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
PyMySQL==1.1.0
python-dotenv==0.21.1
redis==5.0.1
SQLAlchemy==2.0.23
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
Flask-JWT-Extended==4.3.1

